perf: split AI chat from ordinary search loading
