+++ src/contexts/SearchContext.tsx
@@ -56,7 +56,6 @@ export function SearchProvider({ children }: { children: React.ReactNode }) {
 
   const openAiDock = React.useCallback(() => {
     cancelAiDockHoverClose()
-    setHasLoadedSearch(true)
     setIsOpen(false)
     setIsAiDockOpen(true)
   }, [cancelAiDockHoverClose])
@@ -69,7 +68,6 @@ export function SearchProvider({ children }: { children: React.ReactNode }) {
       }
 
       cancelAiDockHoverClose()
-      setHasLoadedSearch(true)
       setIsOpen(false)
       setIsAiDockDirty(true)
       setIsAiDockOpen(true)
